+++ Runtime/Helpers/EnumHelper.cs
@@ -46,5 +46,29 @@ namespace FishingCactus
 
             return ( T )array.GetValue( UnityEngine.Random.Range( min_index, max_index + 1 ) );
         }
+
+        public static T Clamp<T>(
+            T value,
+            T min,
+            T max
+            ) where T : IComparable
+        {
+            T clamped;
+
+            if( value.CompareTo( min ) < 0 )
+            {
+                clamped = min;
+            }
+            else if( value.CompareTo( max ) > 0 )
+            {
+                clamped = max;
+            }
+            else
+            {
+                clamped = value;
+            }
+
+            return clamped;
+        }
     }
 }